Add files via upload
The Course Site Generator application intends to automate the process of building and updating a
course Web site in one easy to use tool. The sites produced by this application will look good and will
be customizable in a number of different ways, but will exist within a common site and page structure.